+++ WhistleWind.AbnormalSigils/Abilities/Special Abilities/MiniGiantCard.cs
@@ -48,18 +48,25 @@ namespace WhistleWind.AbnormalSigils
             if (!info.HasSpecialAbility(MiniGiantCard.Id))
                 return;
 
-            Transform child;
-            if (SaveManager.SaveFile.IsGrimora)
-                child = __result.transform.GetChild(0);
-            else
-                child = __result.transform;
+            Transform child = SaveManager.SaveFile.IsGrimora ? __result.transform.GetChild(0) : __result.transform;
+            if (info.HasTrait(Trait.Giant))
+            {
+                if (SaveManager.SaveFile.IsGrimora)
+                    child.localPosition = new Vector3(0.3f, 0f, 0f);
+                else
+                    child.localPosition = new Vector3(-0.3f, 0.025f, 0f);
 
-            if (SaveManager.SaveFile.IsGrimora)
-                child.localPosition = new Vector3(-0.7f, 1.05f, 0f);
+                child.localScale = new Vector3(0.485f, 1f, 1f);
+            }
             else
-                child.localPosition = new Vector3(0.7f, 0.025f, 1.05f);
+            {
+                if (SaveManager.SaveFile.IsGrimora)
+                    child.localPosition = new Vector3(-0.7f, 1.05f, 0f);
+                else
+                    child.localPosition = new Vector3(0.7f, 0.025f, 1.05f);
 
-            child.localScale = new Vector3(2.1f, 2.1f, 1f);
+                child.localScale = new Vector3(2.1f, 2.1f, 1f);
+            }
         }
 
         [HarmonyTranspiler, HarmonyPatch(typeof(BoardManager), nameof(BoardManager.AssignCardToSlot), MethodType.Enumerator)]
@@ -88,12 +95,46 @@ namespace WhistleWind.AbnormalSigils
         {
             if (card.Info.HasSpecialAbility(MiniGiantCard.Id))
             {
-                if (SaveManager.SaveFile.IsGrimora)
-                    return new Vector3(-0.7f, 1.05f, 0f);
+                if (card.HasTrait(Trait.Giant))
+                {
+                    if (SaveManager.SaveFile.IsGrimora)
+                        return new Vector3(0.3f, 0f, 0f);
+                    else
+                        return new Vector3(-0.3f, 0.025f, 0f);
+                }
                 else
-                    return new Vector3(0.7f, 0.025f, 1.05f);
+                {
+                    if (SaveManager.SaveFile.IsGrimora)
+                        return new Vector3(-0.7f, 1.05f, 0f);
+                    else
+                        return new Vector3(0.7f, 0.025f, 1.05f);
+                }
             }
             return Vector3.up * (instance.SlotHeightOffset + card.SlotHeightOffset);
+        }
+
+        [HarmonyPrefix, HarmonyPatch(typeof(CardAbilityIcons), nameof(CardAbilityIcons.UpdateAbilityIcons))]
+        private static void FixAbilityIcons(CardAbilityIcons __instance, CardInfo info)
+        {
+            if (info.LacksSpecialAbility(MiniGiantCard.Id) || info.LacksTrait(Trait.Giant))
+                return;
+
+            //__instance.transform.localScale = new(1f, 0.485f, 1f);
+            Transform fourGroup = __instance.transform.Find("DefaultIcons_4Abilities");
+
+            Transform child1 = fourGroup.GetChild(0);
+            child1.localPosition = new(-0.12f, 0.16f, 0f);
+            child1.GetComponent<BoxCollider>().size = new(1.25f, 0.6f, 1f);
+            //child1.localScale = new(0.2f, 0.2f, 1f);
+            Transform child2 = fourGroup.GetChild(1);
+            child2.localPosition = new(0.12f, 0.16f, 0f);
+            //child2.localScale = new(0.2f, 0.2f, 1f);
+            Transform child3 = fourGroup.GetChild(2);
+            child3.localPosition = new(-0.12f, -0.135f, 0f);
+            //child3.localScale = new(0.2f, 0.2f, 1f);
+            Transform child4 = fourGroup.GetChild(3);
+            child4.localPosition = new(0.12f, -0.135f, 0f);
+            //child4.localScale = new(0.2f, 0.2f, 1f);
         }
     }